requirements.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.1.0
celery==4.4.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-jinja==2.10.0
django-redis==5.2.0
djangorestframework==3.13.1
flower==0.9.7
gevent==21.12.0
greenlet==1.1.2
humanize==0.5.1
isort==5.10.1
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
plumbum==1.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.26
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.2
rpyc==5.0.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
tornado==6.1
typing-extensions==4.0.1
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
